dfldfldlfd

+++ admin/remove-member.php
@@ -98,8 +98,8 @@ header('location:../index.php');
                 <td><div class='text-center'>$".$row['amount']."</div></td>
                 <td><div class='text-center'>".$row['services']."</div></td>
                 <td><div class='text-center'>".$row['plan']." Month/s</div></td>
-                <td><div class='text-center'><a href='actions/delete-member.php?id=".$row['user_id']."' style='color:#F66;'><i class='fas fa-trash'></i> Remove</a></div></td>
-                
+                <td><div class='text-center'><a onclick=\"return confirm('Please confirm deletion');\" href='actions/delete-member.php?id=".$row['user_id']."' style='color:#F66;'><i class='fas fa-trash'></i> Remove</a></div></td>
+
               </tbody>";
          $cnt++;   }
             ?>
